% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_promoter_matrix.R
\name{create_promoter_matrix}
\alias{create_promoter_matrix}
\title{A function that creates annotated promoter matrix from all PWM scan hits. To annotate promoters with PWM hits, we need to ensure that hit precision score is abouve specified threshold.}
\usage{
create_promoter_matrix(all_hits, optimal_threshold, promoters)
}
\arguments{
\item{all_hits}{dataframe containing all hits obrained when scanning promoters}

\item{optimal_threshold}{threshold for each scanned PWM, either defined by user, or by find_threshold function}

\item{promoters}{GenomicRanges object containing promoters of all active genes in sample, centered on dominant TSS}
}
\value{
promoter_matrix object that for each promoter annotates
}
\description{
A function that creates annotated promoter matrix from all PWM scan hits. To annotate promoters with PWM hits, we need to ensure that hit precision score is abouve specified threshold.
}
