% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomart_promoters.R
\name{get_biomart_promoters}
\alias{get_biomart_promoters}
\title{Extract promoter regions from biomaRt object}
\usage{
get_biomart_promoters(mart, upstream = 500, downstream = 500,
  output = "GR", protein_coding_only = TRUE)
}
\arguments{
\item{mart}{object of class Mart created using the useMart from BiomaRt package}

\item{upstream}{upstream region from TSS to be included in promoter region}

\item{downstream}{downstream region from TSS to be included in promoter region}

\item{output}{character defining output type. 
use "DF" and "GR" or "data.frame" and "GenomicRanges"}

\item{protein_coding_only}{logical: will returned promoters be promoter of 
             only promoter coding genes 

This function extracts promoter regions from a specified mart object. 
Each promoter will be annotated with gene_biotype, tpm expression of a 
promoter and interquantile width of a promoter.
Width of extracted promoters is defined by upstream and downstream parameters. 
Generated promoter regions can be exported as GenomicRanges object or, 
as a data frame.}
}
\value{
a promoter regions of specified width as a data frame or GenomicRanges object
}
\description{
Extract promoter regions from biomaRt object
}
\keyword{promoters}
