% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_promoters.R
\name{scan_promoters}
\alias{scan_promoters}
\title{Scan promoters with PWM and return hits in a tidy format}
\usage{
scan_promoters(seq, pwm, min_score, save_intermediate_files = FALSE,
  label = NA)
}
\arguments{
\item{seq}{A DNAStringSet that will be scanned with PWM}

\item{pwm}{PWMatrix or PWMatrixList object that sequences are going to be scanned against.}

\item{min_score}{A minimum sequence identity needed to report sequence fragment as hit.
Can be given an character string in the format of "80%" or as a single absolute value between 0 and 1.}

\item{save_intermediate_files}{logical: do you want to save a SiteSet object from sequence scan}

\item{label}{if save_intermediate_files = TRUE, label will be used as a filename}
}
\value{
A data.frame that reports all hits from sequence seach in tidy format.
}
\description{
Scan promoters with PWM and return hits in a tidy format
}
